# api/openapi.yaml
openapi: 3.0.3
info:
  title: 관리자 서비스 API
  version: 0.1.0
  description: |
    ① React와 ② JSP + API가 호출하는 REST API의 정본 (ADR-0020, 명세 우선).
    규약·업무 설명은 docs/06-api-spec.md, 모듈별 설명은 docs/06-api/*.md.
    API를 바꿀 때는 이 파일을 먼저 고친다. 서버는 계약 테스트로, React는 생성 타입으로 맞춘다.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /common/codes/{groupCd}:
    get:
      tags: [common]
      operationId: getCodes
      summary: 그룹코드의 상세코드 목록 (코드 콤보용)
      parameters:
        - name: groupCd
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Z0-9_]+$'
          example: USER_STATUS
        - name: includeUnused
          in: query
          description: Y면 사용 안 함 코드도 포함 (이미 저장된 값 표시용)
          schema:
            $ref: '#/components/schemas/YnFlag'
      responses:
        '200':
          description: 정렬 순서대로
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [user]
      operationId: getUser
      summary: 회원 상세 (USR-02). 개인정보는 화면 마스킹 설정을 따르고, 가린 항목은 maskedFields로 알려 준다
      responses:
        '200':
          description: 회원 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      tags: [user]
      operationId: updateUser
      summary: 회원 정보·소속 기업 수정 (USR-05, 06). UPDATE와 PRIVACY 권한이 모두 필요. 회원 구분·아이디는 무시
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: 수정 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [user]
      operationId: deleteUser
      summary: 회원 삭제 (USR-09). 삭제 표시, 사유 필수. 어느 상태에서든 할 수 있다
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDeleteRequest'
      responses:
        '200':
          description: 삭제 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access Token (docs/04-features/09-auth.md 1.1)
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: 입력값 오류 (VALIDATION_ERROR, INVALID_REQUEST)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 인증 필요 (UNAUTHORIZED, TOKEN_EXPIRED 등)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 권한 없음 (FORBIDDEN, PASSWORD_CHANGE_REQUIRED 등)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 대상 없음 (NOT_FOUND)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 중복·동시 수정·업무 규칙 위반
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: 서버 오류 (INTERNAL_ERROR)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, data, error]
      properties:
        success:
          type: boolean
          enum: [false]
        data:
          type: object
          nullable: true
          enum: [null]
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      required: [code, message, fieldErrors]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: 화면에 그대로 보여 줄 수 있는 문장
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: 요청 필드 이름
        message:
          type: string
    NullError:
      type: object
      nullable: true
      enum: [null]
      description: 성공 응답의 error는 항상 null
...
    YnFlag:
      type: string
      enum: ['Y', 'N']
    DateTime:
      type: string
      description: 한국 시간 일시, 초 단위, 시간대 표기 없음 (docs/06-api-spec.md 1절). 동시 수정 방지의 modDt도 이 형식
      pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
      example: '2026-10-04T14:30:15'

    # 오류 코드 목록. admin-core의 ErrorCode enum과 같아야 한다 (ErrorCodeSpecTest).
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - INVALID_REQUEST
        - UNAUTHORIZED
        - TOKEN_EXPIRED
        - REFRESH_FAILED
        - LOGIN_FAILED
        - FORBIDDEN
        - PASSWORD_CHANGE_REQUIRED
        - ACCOUNT_LOCKED
        - ACCOUNT_DISABLED
        - NOT_FOUND
        - DUPLICATE
        - CONFLICT_MODIFIED
        - EXCEL_LIMIT_EXCEEDED
        - FILE_TOO_LARGE
        - INTERNAL_ERROR
        - INVALID_STATUS_CHANGE
        - REASON_REQUIRED
        - COMPANY_HAS_MEMBERS
        - COMPANY_NOT_ACTIVE
        - COMPANY_REQUIRED
        - USER_WITHDRAWN
        - MENU_DEPTH_EXCEEDED
        - MENU_PARENT_NOT_FOLDER
        - MENU_HAS_CHILDREN
        - MENU_SYSTEM_PROTECTED
        - MENU_BOARD_MANAGED
        - MENU_ORDER_MISMATCH
        - MENU_NOT_PAGE
        - CODE_SYSTEM_PROTECTED
        - CODE_GROUP_HAS_CODES
        - BOARD_MENU_EXISTS
        - BOARD_MENU_NOT_FOUND
        - REPLY_NOT_ALLOWED
        - COMMENT_NOT_ALLOWED
        - COMMENT_DEPTH_EXCEEDED
        - ATTACH_NOT_ALLOWED
        - ATTACH_COUNT_EXCEEDED
        - FILE_TYPE_NOT_ALLOWED
        - ROLE_REQUIRED
        - SELF_ROLE_CHANGE
        - SELF_DISABLE
        - PRIVILEGE_ESCALATION
        - LAST_SUPER_ADMIN
        - ROLE_NOT_USABLE
        - ROLE_NOT_EDITABLE
        - ROLE_IN_USE
        - ROLE_SYSTEM_PROTECTED
    EmptyResponse:
      type: object
      required: [success, data, error]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          nullable: true
          enum: [null]
        error:
          $ref: '#/components/schemas/NullError'

    # ---------------------------------------------------------------
    # 인증 (docs/06-api/00-auth.md)
    # ---------------------------------------------------------------
...
    CodeItem:
      type: object
      required: [code, codeNm, sortOrd, useYn]
      properties:
        code:
          type: string
          example: SUSPENDED
        codeNm:
          type: string
          example: 정지
        sortOrd:
          type: integer
          minimum: 0
        useYn:
          $ref: '#/components/schemas/YnFlag'

    # 성공 응답은 { success: true, data, error: null } 틀을 응답마다 펼쳐서 쓴다.
    # (allOf로 조합하면 계약 테스트의 "정의되지 않은 필드" 검사와 맞지 않는다)
    CodeListResponse:
      type: object
      required: [success, data, error]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: '#/components/schemas/CodeItem'
        error:
          $ref: '#/components/schemas/NullError'
...
    PrivacyField:
      type: string
      description: 개인정보 항목 (코드 그룹 PRIVACY_FIELD)
      enum: [USER_NM, EMAIL, MOBILE_NO, BIRTH_DATE]
...
    UserDetail:
      type: object
      required: [userId, userTypeCd, userTypeNm, loginId, userNm, email, mobileNo, companyId, companyNm, statusCd, statusNm, joinDt, birthDate, deptNm, positionNm, joinPath, pwdTempYn, lastLoginDt, withdrawDt, maskedFields, regNm, regDt, modNm, modDt]
      properties:
        userId: { type: integer, format: int64 }
        userTypeCd: { type: string, enum: [PERSONAL, CORPORATE] }
        userTypeNm: { type: string, nullable: true }
        loginId: { type: string }
        userNm: { type: string, description: 화면 마스킹 설정을 따른다 }
        email: { type: string, description: 화면 마스킹 설정을 따른다 }
        mobileNo: { type: string, nullable: true, description: '숫자만, 마스킹하면 010-****-5678' }
        companyId: { type: integer, format: int64, nullable: true }
        companyNm: { type: string, nullable: true }
        statusCd: { type: string, enum: [ACTIVE, DORMANT, SUSPENDED, WITHDRAWN] }
        statusNm: { type: string, nullable: true }
        joinDt: { $ref: '#/components/schemas/DateTime' }
        birthDate: { type: string, nullable: true, description: '원문 2026-10-04, 마스킹하면 1990-**-**' }
        deptNm: { type: string, nullable: true }
        positionNm: { type: string, nullable: true }
        joinPath: { type: string, enum: [USER_SERVICE, ADMIN], description: 'USER_SERVICE 사용자 서비스 가입, ADMIN 관리자 등록' }
        pwdTempYn: { $ref: '#/components/schemas/YnFlag' }
        lastLoginDt: { type: string, nullable: true, pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$' }
        withdrawDt: { type: string, nullable: true, pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$' }
        maskedFields:
          type: array
          description: 가려서 준 항목. 비어 있지 않고 PRIVACY 권한이 있으면 [원문 보기]를 쓸 수 있다
          items: { $ref: '#/components/schemas/PrivacyField' }
        regNm: { type: string, nullable: true }
        regDt: { $ref: '#/components/schemas/DateTime' }
        modNm: { type: string, nullable: true }
        modDt: { $ref: '#/components/schemas/DateTime' }
    UserDetailResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: '#/components/schemas/UserDetail' }
        error: { $ref: '#/components/schemas/NullError' }
...
    UserRequest:
      type: object
      required: [userNm, email]
      additionalProperties: false
      description: 등록은 userTypeCd·loginId 필수·modDt 없음, 수정은 modDt 필수·userTypeCd·loginId 무시
      properties:
        userTypeCd: { type: string, nullable: true, enum: [PERSONAL, CORPORATE, null] }
        loginId: { type: string, nullable: true, pattern: '^[a-z0-9]{4,50}$' }
        userNm: { type: string, minLength: 1, maxLength: 50 }
        email: { type: string, maxLength: 100 }
        mobileNo: { type: string, nullable: true, pattern: '^[0-9]{10,11}$' }
        birthDate: { type: string, format: date, nullable: true }
        companyId: { type: integer, format: int64, nullable: true }
        deptNm: { type: string, nullable: true, maxLength: 100 }
        positionNm: { type: string, nullable: true, maxLength: 50 }
        modDt: { type: string, nullable: true, pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$' }
...
    UserDeleteRequest:
      type: object
      required: [reason]
      additionalProperties: false
      properties:
        reason: { type: string, minLength: 1, maxLength: 500 }
...